Move ProductHint and CoproductHint to generic
